refactor(threads): move thread classes to thread.threads
