+++ src/pamiq_core/thread/threads/__init__.py
@@ -1,0 +1,12 @@
+from .base import BackgroundThread, Thread
+from .control import ControlThread
+from .inference import InferenceThread
+from .training import TrainingThread
+
+__all__ = [
+    "Thread",
+    "BackgroundThread",
+    "ControlThread",
+    "InferenceThread",
+    "TrainingThread",
+]
